+++ src/onegov.activity/onegov/activity/models/__init__.py
@@ -3,6 +3,7 @@ from onegov.activity.models.attendee import Attendee
 from onegov.activity.models.booking import Booking
 from onegov.activity.models.invoice import Invoice
 from onegov.activity.models.invoice_item import InvoiceItem
+from onegov.activity.models.invoice_reference import InvoiceReference
 from onegov.activity.models.occasion import Occasion
 from onegov.activity.models.occasion_date import OccasionDate, DAYS
 from onegov.activity.models.period import Period
@@ -14,6 +15,7 @@ __all__ = [
     'Booking',
     'Invoice',
     'InvoiceItem',
+    'InvoiceReference',
     'Occasion',
     'OccasionDate',
     'Period',